% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikert-package.R
\docType{package}
\name{llikert-package}
\alias{llikert}
\alias{llikert-package}
\title{llikert: Likert scales and Multiple Choice with Next-Token Probabilities}
\description{
Reproducibly rate texts on ordinal or categorical scales using next-token probabilities from an open-weights language model with llama.cpp. Supports local inference through Python and remote inference through Hugging Face Inference Endpoints.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rimonim/llikert}
  \item Report bugs at \url{https://github.com/rimonim/llikert/issues}
}

}
\author{
\strong{Maintainer}: Louis Teitelbaum \email{louis@teitelbaum.us}

Authors:
\itemize{
  \item Louis Teitelbaum \email{louis@teitelbaum.us}
}

}
\keyword{internal}
