% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{write_task}
\alias{write_task}
\alias{read_task}
\title{Save and load tasks as portable JSON}
\usage{
write_task(task, path)

read_task(path)
}
\arguments{
\item{task}{An \code{llikert_task}.}

\item{path}{File path.}
}
\value{
\code{write_task()} returns \code{path} invisibly; \code{read_task()} returns an \code{llikert_task}.
}
\description{
Save and load tasks as portable JSON
}
